Remove GCM and replace it with FCM
