+++ gradle.properties
@@ -1,4 +1,4 @@
-VERSION_NAME=5.4.6-SNAPSHOT
+VERSION_NAME=5.5.0-SNAPSHOT
 
 POM_PACKAGING=aar
 GROUP=com.mixpanel.android
