+++ src/main/java/com/mixpanel/android/mpmetrics/MixpanelAPI.java
@@ -99,8 +99,7 @@ import java.util.concurrent.locks.ReentrantLock;
  * <p>In addition to this documentation, you may wish to take a look at
  * <a href="https://github.com/mixpanel/sample-android-mixpanel-integration">the Mixpanel sample Android application</a>.
  * It demonstrates a variety of techniques, including
- * updating People Analytics records with {@link People} and registering for
- * and receiving push notifications with {@link People#initPushHandling(String)}.
+ * updating People Analytics records with {@link People} and others.
  *
  * <p>There are also <a href="https://mixpanel.com/docs/">step-by-step getting started documents</a>
  * available at mixpanel.com
@@ -421,22 +420,6 @@ public class MixpanelAPI {
 
             return instance;
         }
-    }
-
-    /**
-     * This call is a no-op, and will be removed in future versions.
-     *
-     * @deprecated in 4.0.0, use com.mixpanel.android.MPConfig.FlushInterval application metadata instead
-     */
-    @Deprecated
-    public static void setFlushInterval(Context context, long milliseconds) {
-        MPLog.i(
-            LOGTAG,
-            "MixpanelAPI.setFlushInterval is deprecated. Calling is now a no-op.\n" +
-            "    To set a custom Mixpanel flush interval for your application, add\n" +
-            "    <meta-data android:name=\"com.mixpanel.android.MPConfig.FlushInterval\" android:value=\"YOUR_INTERVAL\" />\n" +
-            "    to the <application> section of your AndroidManifest.xml."
-        );
     }
 
     /**
@@ -811,7 +794,7 @@ public class MixpanelAPI {
      * People Analytics properties.
      *
      * @return an instance of {@link People} that you can use to update
-     *     records in Mixpanel People Analytics and manage Mixpanel Google Cloud Messaging notifications.
+     *     records in Mixpanel People Analytics and manage Mixpanel Firebase Cloud Messaging notifications.
      */
     public People getPeople() {
         return mPeople;
@@ -948,7 +931,6 @@ public class MixpanelAPI {
      *      public void onCreate(Bundle saved) {
      *          mMixpanel = MixpanelAPI.getInstance(this, "YOUR MIXPANEL API TOKEN");
      *          mMixpanel.getPeople().identify("A UNIQUE ID FOR THIS USER");
-     *          mMixpanel.getPeople().initPushHandling("YOUR 12 DIGIT GOOGLE SENDER API");
      *          ...
      *      }
      *
@@ -971,7 +953,7 @@ public class MixpanelAPI {
     public interface People {
         /**
          * Associate future calls to {@link #set(JSONObject)}, {@link #increment(Map)},
-         * and {@link #initPushHandling(String)} with a particular People Analytics user.
+         * {@link #append(String, Object)}, etc... with a particular People Analytics user.
          *
          * <p>All future calls to the People object will rely on this value to assign
          * and increment properties. The user identification will persist across
@@ -1147,111 +1129,102 @@ public class MixpanelAPI {
         public boolean isIdentified();
 
         /**
-         * Enable end-to-end Google Cloud Messaging (GCM) from Mixpanel.
+         * This is a no-op, and will be removed in future versions of the library.
          *
-         * <p>Calling this method will allow the Mixpanel libraries to handle GCM user
-         * registration, and enable Mixpanel to show alerts when GCM messages arrive.
-         *
-         * <p>To use {@link People#initPushHandling}, you will need to add the following to your application manifest:
+         * @deprecated in 5.5.0. Google Cloud Messaging (GCM) is now deprecated by Google.
+         * To enable end-to-end Firebase Cloud Messaging (FCM) from Mixpanel you only need to add
+         * the following to your application manifest XML file:
          *
          * <pre>
          * {@code
-         * <receiver android:name="com.mixpanel.android.mpmetrics.GCMReceiver"
-         *           android:permission="com.google.android.c2dm.permission.SEND" >
-         *     <intent-filter>
-         *         <action android:name="com.google.android.c2dm.intent.RECEIVE" />
-         *         <action android:name="com.google.android.c2dm.intent.REGISTRATION" />
-         *         <category android:name="YOUR_PACKAGE_NAME" />
-         *     </intent-filter>
-         * </receiver>
+         * <service
+         *       android:name="com.mixpanel.android.mpmetrics.MixpanelFCMMessagingService"
+         *       android:enabled="true"
+         *       android:exported="false">
+         *       <intent-filter>
+         *           <action android:name="com.google.firebase.MESSAGING_EVENT"/>
+         *       </intent-filter>
+         * </service>
          * }
          * </pre>
          *
-         * Be sure to replace "YOUR_PACKAGE_NAME" with the name of your package. For
-         * more information and a list of necessary permissions, see {@link GCMReceiver}.
+         * Make sure you have a valid google-services.json file in your project and firebase
+         * messaging is included as a dependency. Example:
          *
-         * <p>If you're planning to use end-to-end support for Messaging, we recommend you
-         * call this method immediately after calling {@link People#identify(String)}, likely
-         * early in your application's life cycle. (for example, in the onCreate method of your
-         * main application activity.)
+         * <pre>
+         * {@code
+         * buildscript {
+         *      ...
+         *      dependencies {
+         *          classpath 'com.google.gms:google-services:3.0.0'
+         *          ...
+         *      }
+         * }
          *
-         * <p>Calls to {@link People#initPushHandling} should not be mixed with calls to
-         * {@link #setPushRegistrationId(String)} and {@link #clearPushRegistrationId()}
-         * in the same application. Application authors should choose one or the other
-         * method for handling Mixpanel GCM messages.
+         * dependencies {
+         *     implementation 'com.google.firebase:firebase-messaging:17.3.4'
+         *     implementation 'com.mixpanel.android:mixpanel-android:5.5.0'
+         * }
          *
-         * @param senderID of the Google API Project that registered for Google Cloud Messaging
-         *     You can find your ID by looking at the URL of in your Google API Console
-         *     at https://code.google.com/apis/console/; it is the twelve digit number after
-         *     after "#project:" in the URL address bar on console pages.
+         * apply plugin: 'com.google.gms.google-services'
+         * }
+         * </pre>
          *
-         * @see com.mixpanel.android.mpmetrics.GCMReceiver
+         * We recommend you update your Server Key on mixpanel.com from your Firebase console. Legacy
+         * server keys are still supported.
+         *
          * @see <a href="https://mixpanel.com/docs/people-analytics/android-push">Getting Started with Android Push Notifications</a>
          */
+        @Deprecated
         public void initPushHandling(String senderID);
 
         /**
-         * Retrieves Google Cloud Messaging registration ID.
+         * Retrieves current Firebase Cloud Messaging token.
          *
          * <p>{@link People#getPushRegistrationId} should only be called after {@link #identify(String)} has been called.
          *
-         * @return GCM push token or null if the user has not been registered in GCM.
+         * @return FCM push token or null if the user has not been registered in FCM.
          *
-         * @see #initPushHandling(String)
          * @see #setPushRegistrationId(String)
+         * @see #clearPushRegistrationId
          */
         public String getPushRegistrationId();
 
         /**
-         * Manually send a Google Cloud Messaging registration id to Mixpanel.
+         * Manually send a Firebase Cloud Messaging token to Mixpanel.
          *
-         * <p>If you are handling Google Cloud Messages in your own application, but would like to
+         * <p>If you are handling Firebase Cloud Messages in your own application, but would like to
          * allow Mixpanel to handle messages originating from Mixpanel campaigns, you should
-         * call setPushRegistrationId with the "registration_id" property of the
-         * com.google.android.c2dm.intent.REGISTRATION intent when it is received.
+         * call setPushRegistrationId with the FCM token.
          *
          * <p>setPushRegistrationId should only be called after {@link #identify(String)} has been called.
          *
-         * <p>Calls to {@link People#setPushRegistrationId} should not be mixed with calls to {@link #initPushHandling(String)}
-         * in the same application. In addition, applications that call setPushRegistrationId
-         * should also call {@link #clearPushRegistrationId()} when they receive an intent to unregister
-         * (a com.google.android.c2dm.intent.REGISTRATION intent with getStringExtra("unregistered") != null)
+         * Optionally, applications that call setPushRegistrationId should also call
+         * {@link #clearPushRegistrationId()} when they unregister the device id.
          *
-         * @param registrationId the result of calling intent.getStringExtra("registration_id")
-         *     on a com.google.android.c2dm.intent.REGISTRATION intent
+         * @param token Firebase Cloud Messaging token
          *
-         * @see #initPushHandling(String)
          * @see #clearPushRegistrationId()
          */
-        public void setPushRegistrationId(String registrationId);
+        public void setPushRegistrationId(String token);
 
         /**
-         * Manually clear all current Google Cloud Messaging registration ids from Mixpanel.
-         *
-         * <p>If you are handling Google Cloud Messages in your own application, you should
-         * call this method when your application receives a com.google.android.c2dm.intent.REGISTRATION
-         * with getStringExtra("unregistered") != null
+         * Manually clear all current Firebase Cloud Messaging tokens from Mixpanel.
          *
          * <p>{@link People#clearPushRegistrationId} should only be called after {@link #identify(String)} has been called.
          *
          * <p>In general, all applications that call {@link #setPushRegistrationId(String)} should include a call to
-         * clearPushRegistrationId, and no applications that call
-         * {@link #initPushHandling(String)} should call clearPushRegistrationId
+         * clearPushRegistrationId.
          */
         public void clearPushRegistrationId();
 
         /**
-         * Manually clear a single Google Cloud Messaging registration id from Mixpanel.
-         *
-         * <p>If you are handling Google Cloud Messages in your own application, you should
-         * call this method when your application receives a com.google.android.c2dm.intent.REGISTRATION
-         * with getStringExtra("unregistered") != null
+         * Manually clear a single Firebase Cloud Messaging token from Mixpanel.
          *
          * <p>{@link People#clearPushRegistrationId} should only be called after {@link #identify(String)} has been called.
          *
          * <p>In general, all applications that call {@link #setPushRegistrationId(String)} should include a call to
-         * clearPushRegistrationId, and no applications that call
-         * {@link #initPushHandling(String)} should call clearPushRegistrationId
+         * clearPushRegistrationId.
          */
         public void clearPushRegistrationId(String registrationId);
 
@@ -1428,29 +1401,6 @@ public class MixpanelAPI {
     }
 
     /**
-     * This method is a no-op, kept for compatibility purposes.
-     *
-     * To enable verbose logging about communication with Mixpanel, add
-     * {@code
-     * <meta-data android:name="com.mixpanel.android.MPConfig.EnableDebugLogging" />
-     * }
-     *
-     * To the {@code <application>} tag of your AndroidManifest.xml file.
-     *
-     * @deprecated in 4.1.0, use Manifest meta-data instead
-     */
-    @Deprecated
-    public void logPosts() {
-        MPLog.i(
-                LOGTAG,
-                "MixpanelAPI.logPosts() is deprecated.\n" +
-                        "    To get verbose debug level logging, add\n" +
-                        "    <meta-data android:name=\"com.mixpanel.android.MPConfig.EnableDebugLogging\" value=\"true\" />\n" +
-                        "    to the <application> section of your AndroidManifest.xml."
-        );
-    }
-
-    /**
      * Attempt to register MixpanelActivityLifecycleCallbacks to the application's event lifecycle.
      * Once registered, we can automatically check for and show in-app notifications
      * when any Activity is opened.
@@ -1503,7 +1453,7 @@ public class MixpanelAPI {
         mSessionMetadata.initSession();
     }
 
-    // Package-level access. Used (at least) by GCMReceiver
+    // Package-level access. Used (at least) by MixpanelFCMMessagingService
     // when OS-level events occur.
     /* package */ interface InstanceProcessor {
         public void process(MixpanelAPI m);
@@ -1920,28 +1870,33 @@ public class MixpanelAPI {
 
         @Override
         public void initPushHandling(String senderID) {
-            if (! ConfigurationChecker.checkPushConfiguration(mContext) ) {
-                MPLog.i(LOGTAG, "Can't register for push notification services. Push notifications will not work.");
-                MPLog.i(LOGTAG, "See log tagged " + ConfigurationChecker.LOGTAG + " above for details.");
-            }
-            else { // Configuration is good for at least some push notifications
-                final String pushId = mPersistentIdentity.getPushId();
-                if (pushId == null) {
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                        registerForPushIdAPI21AndUp(senderID);
-                    } else {
-                        registerForPushIdAPI19AndOlder(senderID);
-                    }
-                } else {
-                    MixpanelAPI.allInstances(new InstanceProcessor() {
-                        @Override
-                        public void process(MixpanelAPI api) {
-                            MPLog.v(LOGTAG, "Using existing pushId " + pushId);
-                            api.getPeople().setPushRegistrationId(pushId);
-                        }
-                    });
-                }
-            }// endelse
+            MPLog.w(
+                    LOGTAG,
+                    "MixpanelAPI.initPushHandling is deprecated. This is a no-op.\n" +
+                            "   Mixpanel now uses Firebase Cloud Messaging. You need to remove your old Mixpanel" +
+                            " GCM Receiver from your AndroidManifest.xml and add the following:\n" +
+                            "   <service\n" +
+                            "       android:name=\"com.mixpanel.android.mpmetrics.MixpanelFCMMessagingService\"\n" +
+                            "       android:enabled=\"true\"\n" +
+                            "       android:exported=\"false\">\n" +
+                            "       <intent-filter>\n" +
+                            "           <action android:name=\"com.google.firebase.MESSAGING_EVENT\"/>\n" +
+                            "       </intent-filter>\n" +
+                            "   </service>\n\n" +
+                            "Make sure to add firebase messaging as a dependency on your gradle file:\n" +
+                            "buildscript {\n" +
+                            "   ...\n" +
+                            "   dependencies {\n" +
+                            "       classpath 'com.google.gms:google-services:3.0.0'\n" +
+                            "       ...\n" +
+                            "   }\n" +
+                            "}\n" +
+                            "dependencies {\n" +
+                            "   implementation 'com.google.firebase:firebase-messaging:17.3.4'\n" +
+                            "   implementation 'com.mixpanel.android:mixpanel-android:5.5.0'\n" +
+                            "}\n" +
+                            "apply plugin: 'com.google.gms.google-services'"
+            );
         }
 
         @Override
@@ -2010,24 +1965,6 @@ public class MixpanelAPI {
             dataObj.put("$mp_metadata", mSessionMetadata.getMetadataForPeople());
 
             return dataObj;
-        }
-
-        @TargetApi(21)
-        private void registerForPushIdAPI21AndUp(String senderID) {
-            mMessages.registerForGCM(senderID);
-        }
-
-        @TargetApi(19)
-        private void registerForPushIdAPI19AndOlder(String senderID) {
-            try {
-                MPLog.v(LOGTAG, "Registering a new push id");
-                final Intent registrationIntent = new Intent("com.google.android.c2dm.intent.REGISTER");
-                registrationIntent.putExtra("app", PendingIntent.getBroadcast(mContext, 0, new Intent(), 0));
-                registrationIntent.putExtra("sender", senderID);
-                mContext.startService(registrationIntent);
-            } catch (final SecurityException e) {
-                MPLog.w(LOGTAG, "Error registering for push notifications", e);
-            }
         }
 
         private void showGivenOrAvailableNotification(final InAppNotification notifOrNull, final Activity parent) {
